+++ bin/check-versions.js
@@ -3,6 +3,7 @@ var chalk = require('chalk')
 var semver = require('semver')
 var packageConfig = require(path.resolve(process.cwd(), 'package.json'))
 var shell = require('shelljs')
+
 function exec (cmd) {
   return require('child_process').execSync(cmd).toString().trim()
 }