+++ bin/dll.js
@@ -4,10 +4,11 @@ var chalk = require('chalk')
 var config = require('../config')
 
 if (!config.dll.enable) {
-  console.log(chalk.yellow('如果要执行dll，请在custom.config.js的dll配置中设置enable:true,并设置相应的entry\n'))
+  console.log(chalk.yellow('Enalbed the dll config'))
   process.exit(0)
 }
-console.log(chalk.cyan('building dll...\n'))
+
+console.log(chalk.cyan('Building dll...\n'))
 
 var dllConf = require('./webpack.dll.conf')
 var compiler = webpack(dllConf)
@@ -21,5 +22,5 @@ compiler.run(function (err, stats) {
       chunks: false,
       chunkModules: false
     }) + '\n')
-  console.log(chalk.cyan('build dll complete.\n'))
+  console.log(chalk.cyan('Build dll complete.\n'))
 })