+++ bin/webpack.dev.conf.js
@@ -1,7 +1,8 @@
 var path = require('path')
-var utils = require('../utils')
+var chalk = require('chalk')
 var webpack = require('webpack')
 var config = require('../config')
+var utils = require('../utils')
 var merge = require('webpack-merge')
 var baseWebpackConfig = require('./webpack.base.conf')
 var FriendlyErrorsPlugin = require('friendly-errors-webpack-plugin')
@@ -11,12 +12,7 @@ Object.keys(baseWebpackConfig.entry).forEach(function (name) {
   baseWebpackConfig.entry[ name ] = [ path.resolve(__dirname, './dev-client') ].concat(baseWebpackConfig.entry[ name ])
 })
 var plugins = [
-  new webpack.DefinePlugin({
-    'process.env': config.dev.env
-  }),
-  // https://github.com/glenjamin/webpack-hot-middleware#installation--usage
   new webpack.HotModuleReplacementPlugin(),
-  new webpack.NoEmitOnErrorsPlugin(),
   new FriendlyErrorsPlugin()
 ]
 
@@ -30,10 +26,16 @@ if (config.dll.enable) {
 }
 
 module.exports = merge(baseWebpackConfig, {
+  mode: 'development',
   module: {
     rules: utils.styleLoaders({ sourceMap: config.dev.cssSourceMap })
   },
-  // cheap-module-eval-source-map is faster for development
-  devtool: '#cheap-module-eval-source-map',
+  optimization: {
+    runtimeChunk: false,
+    minimize: false,
+    noEmitOnErrors: true,
+    splitChunks: false
+  },
+  devtool: config.dev.devtool,
   plugins: plugins
-})
+}, config.custom.dev)